hooks up search
