+++ src/App.css
@@ -54,3 +54,9 @@ body {
   transition: width 0.3s linear;
   width: var(--right-panel-width);
 }
+
+.main-search {
+  padding-bottom: 1rem;
+  width: 14rem;
+  margin-left: auto;
+}
